# Use ETS table to speed up HashRing via process read operations