--- a/lib/app.ex
+++ b/lib/app.ex
@@ -18,10 +18,11 @@ defmodule HashRing.App do
     Enum.each(Application.get_env(:libring, :rings, []), fn
       {name, config} ->
         {:ok, _pid} = HashRing.Managed.new(name, config)
-        Logger.info "[libring] started managed ring #{inspect name}"
+        Logger.info("[libring] started managed ring #{inspect(name)}")
+
       name when is_atom(name) ->
         {:ok, _pid} = HashRing.Managed.new(name)
-        Logger.info "[libring] started managed ring #{inspect name}"
+        Logger.info("[libring] started managed ring #{inspect(name)}")
     end)
 
     # Application started