--- a/lib/utils.ex
+++ b/lib/utils.ex
@@ -2,7 +2,7 @@ defmodule HashRing.Utils do
   @moduledoc false
   require Logger
 
-  @type pattern_list :: [String.t | Regex.t]
+  @type pattern_list :: [String.t() | Regex.t()]
   @type blacklist :: pattern_list
   @type whitelist :: pattern_list
 
@@ -20,56 +20,76 @@ defmodule HashRing.Utils do
 
   def ignore_node?(_node, [], []),
     do: false
+
   def ignore_node?(node, blacklist, whitelist) when is_atom(node),
     do: ignore_node?(Atom.to_string(node), blacklist, whitelist)
+
   def ignore_node?(node, blacklist, []) when is_binary(node) and is_list(blacklist) do
     Enum.any?(blacklist, fn
       ^node ->
         true
+
       %Regex{} = pattern ->
         Regex.match?(pattern, node)
+
       pattern when is_binary(pattern) ->
         case Regex.compile(pattern) do
           {:ok, rx} ->
             Regex.match?(rx, node)
+
           {:error, reason} ->
-            :ok = Logger.warn "[libring] ignore_node?/3: invalid blacklist pattern (#{inspect pattern}): #{inspect reason}"
+            :ok =
+              Logger.warn(
+                "[libring] ignore_node?/3: invalid blacklist pattern (#{inspect(pattern)}): #{inspect(reason)}"
+              )
+
             false
         end
     end)
   end
+
   def ignore_node?(node, [], whitelist) when is_binary(node) and is_list(whitelist) do
     not Enum.any?(whitelist, fn
       ^node ->
         true
+
       %Regex{} = pattern ->
         Regex.match?(pattern, node)
+
       pattern when is_binary(pattern) ->
         case Regex.compile(pattern) do
           {:ok, rx} ->
             Regex.match?(rx, node)
+
           {:error, reason} ->
-            :ok = Logger.warn "[libring] ignore_node?/3: invalid whitelist pattern (#{inspect pattern}): #{inspect reason}"
+            :ok =
+              Logger.warn(
+                "[libring] ignore_node?/3: invalid whitelist pattern (#{inspect(pattern)}): #{inspect(reason)}"
+              )
+
             false
         end
     end)
   end
+
   def ignore_node?(node, blacklist, whitelist) when is_list(whitelist) and is_list(blacklist) do
     # Criteria for ignoring nodes when both blacklisting and whitelisting is active
-    blacklisted? =
-      ignore_node?(node, blacklist, [])
-    whitelisted? =
-      not ignore_node?(node, [], whitelist)
+    blacklisted? = ignore_node?(node, blacklist, [])
+    whitelisted? = not ignore_node?(node, [], whitelist)
+
     cond do
       # If it is blacklisted and also whitelisted, then do not ignore
       blacklisted? and whitelisted? ->
         false
+
       # If it is blacklisted and not also whitelisted, then ignore
       blacklisted? ->
         true
+
       # If it is not blacklisted and is whitelisted, then do not ignore
       whitelisted? ->
         false
+
       # If it is not blacklisted and not whitelisted, then ignore
       :else ->
         true