--- a/mix.exs
+++ b/mix.exs
@@ -21,8 +21,7 @@ defmodule HashRing.Mixfile do
       preferred_cli_env: [
         docs: :docs,
         dialyzer: :test,
-        "hex.publish": :docs,
-        "eqc.mini": :test
+        "hex.publish": :docs
       ]
     ]
   end
@@ -42,8 +41,8 @@ defmodule HashRing.Mixfile do
 
       {:ex_doc, ">= 0.0.0", only: [:docs]},
       {:benchee, "~> 1.0", only: [:dev]},
-      {:dialyxir, "~> 1.0", only: [:test]},
-      {:eqc_ex, "~> 1.4", only: [:test]}
+      {:dialyxir, "~> 1.0", only: [:test], runtime: false},
+      {:stream_data, "~> 0.5", only: [:test]}
     ]
   end