… config/config.exs
… mix.lock
… test/hashring_utils_test.exs