Before:
```
Elixir 1.12.2
Erlang 23.3.4.5

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 10 s
memory time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 24 s

Benchmarking HashRing.key_to_node (direct)...
Benchmarking HashRings.key_to_node (via process)...

Name                                          ips        average  deviation         median         99th %
HashRing.key_to_node (direct)             24.68 K       40.52 μs   ±397.13%       34.98 μs       94.98 μs
HashRings.key_to_node (via process)        4.20 K      238.22 μs    ±23.01%      210.98 μs      445.98 μs
```

After:
```
Elixir 1.12.2
Erlang 23.3.4.5

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 10 s
memory time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 24 s

Benchmarking HashRing.key_to_node (direct)...
Benchmarking HashRings.key_to_node (via process)...

Name                                          ips        average  deviation         median         99th %
HashRing.key_to_node (direct)             24.18 K       41.35 μs    ±21.89%          39 μs          75 μs
HashRings.key_to_node (via process)       13.58 K       73.66 μs    ±40.90%          61 μs         198 μs
```

Not to mention this benchmark doesn't take into account multicore is being used now.

Anyway, the speedup is ~3.2x.