@edescourtis 2021-08-05
I cannot verify tests are working:
```
== Compilation error in file test/hashring_test.exs ==
** (CompileError) test/hashring_test.exs:3: module :eqc_gen is not loaded and could not be found
    (eqc_ex 1.4.2) expanding macro: EQC.__using__/1
    test/hashring_test.exs:3: HashRingTest (module)
    (elixir 1.12.2) expanding macro: Kernel.use/1
    test/hashring_test.exs:3: HashRingTest (module)
    (eqc_ex 1.4.2) expanding macro: EQC.ExUnit.__using__/1
    test/hashring_test.exs:3: HashRingTest (module)
    (elixir 1.12.2) expanding macro: Kernel.use/1
    test/hashring_test.exs:3: HashRingTest (module)
    (elixir 1.12.2) lib/kernel/parallel_compiler.ex:428: Kernel.ParallelCompiler.require_file/2
    (elixir 1.12.2) lib/kernel/parallel_compiler.ex:321: anonymous fn/4 in Kernel.ParallelCompiler.spawn_workers/7
```

Would you mind verifying on your end?